update run once script

+++ home/.chezmoiscripts/darwin/run_once_before_00-install-prerequisits.sh
@@ -10,11 +10,26 @@ else
     /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 fi
 
+if command -v dcli &> /dev/null; then
+    echo "Dashlane CLI is already installed."
+else
+    # Install Dashlane CLI
+    echo "Install and setup of Dashlane CLI"
+    brew install dashlane/tap/dashlane-cli
+    dcli sync
+fi
+
 # install EurKey Keyboard layout
-echo "Downloading EurKEY keyboard layout. (Requires restart by user)"
-wget https://github.com/jonasdiemer/EurKEY-Mac/raw/master/EurKEY.icns
-wget https://github.com/jonasdiemer/EurKEY-Mac/raw/master/EurKEY.keylayout
-sudo mv EurKEY.icns EurKEY.keylayout "/Library/Keyboard Layouts/"
+# check if layout exists
+if [ -f "/Library/Keyboard Layouts/EurKEY.keylayout" ]; then
+    echo "EurKEY keyboard layout is already installed."
+else
+    echo "Downloading EurKEY keyboard layout. (Requires restart by user)"
+    wget https://github.com/jonasdiemer/EurKEY-Mac/raw/master/EurKEY.icns
+    wget https://github.com/jonasdiemer/EurKEY-Mac/raw/master/EurKEY.keylayout
+    sudo mv EurKEY.icns EurKEY.keylayout "/Library/Keyboard Layouts/"
+fi
+
